Move most swift code into Swift module
