+++ compiler/core/src/color.re
@@ -21,37 +21,3 @@ let find = (colors: list(t), id: string) =>
   | color => Some(color)
   | exception Not_found => None
   };
-
-let render = (target, colors) =>
-  switch target {
-  | Types.Swift =>
-    open Ast.Swift;
-    let colorConstantDoc = (color) =>
-      LineEndComment({
-        "comment": color.value,
-        "line":
-          ConstantDeclaration({
-            "modifiers": [AccessLevelModifier(PublicModifier), StaticModifier],
-            "pattern": IdentifierPattern({"identifier": color.id, "annotation": None}),
-            "init": Some(LiteralExpression(Color(color.value)))
-          })
-      });
-    let doc =
-      TopLevelDeclaration({
-        "statements": [
-          ImportDeclaration("UIKit"),
-          Empty,
-          ClassDeclaration({
-            "name": "Colors",
-            "inherits": [],
-            "modifier": None,
-            "isFinal": false,
-            "body": colors |> List.map(colorConstantDoc)
-          })
-        ]
-      });
-    Render.Swift.toString(doc)
-  | _ =>
-    Js.log2("Color generation not supported for target", target);
-    "error"
-  };
